refactor(mcp-tests): split orchestrator out of mcp_test_base
mcp_test_base.py defined both MCPTestBase (imported by every test
category) and MCPTestOrchestrator (which dynamically imports those
same categories). That made each test_*.py module sit on a cycle:
test_foo → mcp_test_base → tests.test_foo. Python handles the cycle
at runtime because the orchestrator's imports are inside method
bodies, but CodeQL flagged all 44 sites under py/cyclic-import.

Move MCPTestOrchestrator to tests/mcp_test_orchestrator.py. The base
module no longer references any sibling test module, so the cycle is
broken at source. tests/__init__.py keeps re-exporting both names and
run_tests.py points at the new module.
